The way it should have been done.
You may have noticed that the way I interface with the minecraft process
is hacky and terrible. Prone to freezing and breaking and who knows what
else. This is the start of making things right. Interacting with the
Minecraft process the way it's meant to be interacted with.

Also, I'm doing it asyncronously because I need to learn how to use
Python's asyncio library and this seemed like as good a time as any to
learn it.